Split up EventInserter into base and derived classes, ref#142
